requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
pyyaml>=6.0.1
python-socketio>=5.11.0
eventlet>=0.35.2 